# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Termostato API
  description: |
    Contratto REST inbound del backend del termostato intelligente.

    Tutti gli endpoint richiedono l'header `X-API-Key`. La chiave viene
    confrontata con l'elenco `api_keys` della configurazione. Una chiave
    mancante o non valida produce `401 Unauthorized`.

    I nomi JSON seguono la strategia `SNAKE_CASE` configurata nel backend.
    Le date e i timestamp sono espressi in UTC.
    Le ore degli intervalli del calendario sono invece interpretate nel
    `fuso_orario` configurato, con regole DST controllate da `ora_legale`.

    Gli endpoint `/temperature`, `/relay` e `/mock/*` sono disponibili solo
    quando è attivo il profilo Spring `mock`; simulano il sensore e il relay
    per lo scenario E2E. Il client ntfy e gli endpoint configurati per i
    dispositivi fisici sono integrazioni outbound e non fanno parte di questa
    specifica inbound.
  version: 1.2.0
  contact:
    name: Team Termostato
servers:
  - url: http://localhost:8080
    description: Istanza locale predefinita (la porta può essere sovrascritta)
security:
  - ApiKeyAuth: []
...
paths:
  /config:
    get:
      tags: [Configurazione]
      summary: Legge la configurazione corrente
      description: Restituisce la configurazione runtime attualmente applicata.
      operationId: getConfiguration
      responses:
        '200':
          description: Configurazione corrente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfiguration'
              example:
                soglia_attivazione: 0.3
                override_attivo: false
                temperatura_override: null
                override_fine: null
                intervallo_polling_secondi: 60
                max_errori_consecutivi: 3
                retention_log_giorni: 30
                ntfy_url: https://ntfy.sh
                ntfy_topic: sliverd
                debug_mode: false
                fuso_orario: Europe/Rome
                ora_legale: true
                notifiche_errori_abilitate: true
                meteo_esterno_url: https://api.open-meteo.com
                meteo_esterno_latitudine: 37.6167
                meteo_esterno_longitudine: 15.1667
                sensore_url: http://sensore.local
                relay_url: http://relay.local
                database_path: ./data/termostato.db
                api_keys:
                  - sostituire-con-una-chiave-segreta
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Configurazione]
      summary: Aggiorna la configurazione corrente
      description: |
        Sostituisce la configurazione runtime e la persiste nel file JSON.
        L'aggiornamento è immediato e non richiede il riavvio.

        Il campo `database_path` è bootstrap-only e deve rimanere invariato
        rispetto al valore usato all'avvio. La richiesta deve contenere una
        chiave attualmente autorizzata; non è possibile usare questa operazione
        per aggiungere la prima chiave quando `api_keys` è vuoto.

        Se `override_attivo` è true e `override_fine` è valorizzato, deve essere
        un istante futuro rispetto all'ora corrente (orario civile locale):
        un valore già trascorso viene rifiutato con `400`.
      operationId: updateConfiguration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfiguration'
            example:
              soglia_attivazione: 0.3
              override_attivo: true
              temperatura_override: 20.5
              override_fine: '2026-09-13T08:00:00'
              intervallo_polling_secondi: 60
              max_errori_consecutivi: 3
              retention_log_giorni: 30
              ntfy_url: https://ntfy.sh
              ntfy_topic: sliverd
              debug_mode: false
              fuso_orario: Europe/Rome
              ora_legale: true
              notifiche_errori_abilitate: true
              meteo_esterno_url: https://api.open-meteo.com
              meteo_esterno_latitudine: 37.6167
              meteo_esterno_longitudine: 15.1667
              sensore_url: http://sensore.local
              relay_url: http://relay.local
              database_path: ./data/termostato.db
              api_keys:
                - sostituire-con-una-chiave-segreta
      responses:
        '200':
          description: Configurazione aggiornata e persistita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfiguration'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /config/calendario:
    get:
      tags: [Configurazione]
      summary: Legge il calendario settimanale
      description: Restituisce i sette giorni del calendario con gli intervalli configurati.
      operationId: getCalendar
      responses:
        '200':
          description: Calendario corrente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarDocument'
              example:
                giorni:
                  lunedi: []
                  martedi: []
                  mercoledi: []
                  giovedi:
                    - ora_inizio: '06:00'
                      ora_fine: '08:00'
                      temperatura_target: 20.5
                  venerdi: []
                  sabato: []
                  domenica: []
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Configurazione]
      summary: Aggiorna il calendario settimanale
      description: |
        Sostituisce e persiste il calendario. Il payload deve contenere
        esattamente i sette giorni canonici. Gli intervalli dello stesso giorno
        non possono sovrapporsi e `ora_inizio` deve precedere `ora_fine`.
        Gli orari sono interpretati nel `fuso_orario` configurato; quando
        `ora_legale` è true vengono applicate automaticamente le regole CET/CEST.
        La fine dell'intervallo è esclusa.
      operationId: updateCalendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarDocument'
            example:
              giorni:
                lunedi: []
                martedi: []
                mercoledi: []
                giovedi:
                  - ora_inizio: '06:00'
                    ora_fine: '08:00'
                    temperatura_target: 20.5
                venerdi: []
                sabato: []
                domenica: []
      responses:
        '200':
          description: Calendario aggiornato e persistito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarDocument'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API-key autorizzata nella configurazione `api_keys`.
...
  responses:
    Unauthorized:
      description: API-key assente, vuota o non autorizzata
      headers:
        WWW-Authenticate:
          description: Schema di autenticazione richiesto dal backend
          schema:
            type: string
            example: ApiKey
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
          example:
            error: API key mancante o non valida
    BadRequest:
      description: Body o parametri non validi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            timestamp: '2026-09-04T10:00:00Z'
            status: 400
            error: Bad Request
            message: Il parametro da non può essere successivo al parametro a
    InternalServerError:
      description: Errore interno o impossibilità di persistere la modifica
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            timestamp: '2026-09-04T10:00:00Z'
            status: 500
            error: Internal Server Error
            message: Errore interno del server
  schemas:
    SystemConfiguration:
      type: object
      description: |
        Configurazione runtime completa. I valori numerici di temperatura
        devono avere una sola cifra decimale. `fuso_orario` è un identificatore
        IANA; `ora_legale` abilita le regole DST automatiche del fuso. Se
        `override_attivo` è true, `temperatura_override` è obbligatoria; se è false può essere null.
        `override_fine` è opzionale e indica quando termina la forzatura: è un orario civile
        locale (stesso fuso/DST del calendario) e, se presente su `PUT`, deve essere futuro.
      additionalProperties: false
      required:
        - soglia_attivazione
        - override_attivo
        - temperatura_override
        - intervallo_polling_secondi
        - max_errori_consecutivi
        - retention_log_giorni
        - ntfy_url
        - ntfy_topic
        - debug_mode
        - fuso_orario
        - ora_legale
        - notifiche_errori_abilitate
        - meteo_esterno_url
        - meteo_esterno_latitudine
        - meteo_esterno_longitudine
        - sensore_url
        - relay_url
        - database_path
        - api_keys
      properties:
        soglia_attivazione:
          type: number
          format: double
          minimum: 0
          multipleOf: 0.1
          description: Isteresi di attivazione in °C.
          example: 0.3
        override_attivo:
          type: boolean
          description: Abilita l'override manuale del target calendario.
          example: false
        temperatura_override:
          type: number
          format: double
          nullable: true
          multipleOf: 0.1
          description: Target manuale in °C; obbligatorio quando `override_attivo` è true.
          example: 20.5
        override_fine:
          type: string
          format: date-time
          nullable: true
          description: |
            Istante di fine forzatura in orario civile locale (stesso `fuso_orario`/`ora_legale`
            del calendario), formato ISO senza offset (es. `2026-09-13T08:00:00`). Opzionale e
            rilevante solo con `override_attivo` true; se assente la forzatura è a tempo
            indeterminato. Su `PUT` deve essere futuro, altrimenti la richiesta è rifiutata con 400.
            Alla scadenza la forzatura si disattiva automaticamente e viene persistita.
          example: '2026-09-13T08:00:00'
        intervallo_polling_secondi:
          type: integer
          format: int32
          minimum: 1
          description: Intervallo del ciclo di controllo in secondi.
          example: 60
        max_errori_consecutivi:
          type: integer
          format: int32
          minimum: 1
          description: Soglia di errori consecutivi prima della messa in sicurezza.
          example: 3
        retention_log_giorni:
          type: integer
          format: int32
          minimum: 1
          description: Giorni di conservazione dei log.
          example: 30
        ntfy_url:
          type: string
          format: uri
          minLength: 1
          description: URL base del servizio ntfy.
          example: https://ntfy.sh
        ntfy_topic:
          type: string
          minLength: 1
          description: Topic ntfy usato per le notifiche.
          example: sliverd
        debug_mode:
          type: boolean
          description: Abilita le notifiche informative di accensione e spegnimento.
          example: false
        fuso_orario:
          type: string
          minLength: 1
          pattern: '^[A-Za-z0-9._+-]+(?:/[A-Za-z0-9._+-]+)*$'
          description: Identificatore IANA del fuso usato dal calendario.
          default: Europe/Rome
          example: Europe/Rome
        ora_legale:
          type: boolean
          description: Se true applica automaticamente le regole DST del fuso; se false usa l'offset standard.
          default: true
          example: true
        notifiche_errori_abilitate:
          type: boolean
          description: Abilita le notifiche degli errori tramite ntfy.
          default: true
          example: true
        meteo_esterno_url:
          type: string
          format: uri
          minLength: 1
          description: URL base del servizio REST meteo pubblico.
          example: https://api.open-meteo.com
        meteo_esterno_latitudine:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: Latitudine del punto usato per il meteo esterno; default Acireale.
          example: 37.6167
        meteo_esterno_longitudine:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: Longitudine del punto usato per il meteo esterno; default Acireale.
          example: 15.1667
        sensore_url:
          type: string
          format: uri
          minLength: 1
          description: URL base del sensore REST esterno.
          example: http://sensore.local
        relay_url:
          type: string
          format: uri
          minLength: 1
          description: URL base del relay REST esterno.
          example: http://relay.local
        database_path:
          type: string
          minLength: 1
          description: Percorso del database SQLite; bootstrap-only.
          example: ./data/termostato.db
        api_keys:
          type: array
          uniqueItems: true
          description: Elenco delle API-key autorizzate. Lista vuota significa fail-closed.
          items:
            type: string
            minLength: 1
          example:
            - e2e-test-key
    CalendarDocument:
      type: object
      description: Calendario settimanale con esattamente i sette giorni canonici.
      additionalProperties: false
      required: [giorni]
      properties:
        giorni:
          type: object
          additionalProperties: false
          required:
            - lunedi
            - martedi
            - mercoledi
            - giovedi
            - venerdi
            - sabato
            - domenica
          properties:
            lunedi:
              $ref: '#/components/schemas/CalendarIntervals'
            martedi:
              $ref: '#/components/schemas/CalendarIntervals'
            mercoledi:
              $ref: '#/components/schemas/CalendarIntervals'
            giovedi:
              $ref: '#/components/schemas/CalendarIntervals'
            venerdi:
              $ref: '#/components/schemas/CalendarIntervals'
            sabato:
              $ref: '#/components/schemas/CalendarIntervals'
            domenica:
              $ref: '#/components/schemas/CalendarIntervals'
    CalendarIntervals:
      type: array
      description: Intervalli ordinati automaticamente per ora di inizio e non sovrapposti.
      items:
        $ref: '#/components/schemas/TimeInterval'
    TimeInterval:
      type: object
      additionalProperties: false
      required:
        - ora_inizio
        - ora_fine
        - temperatura_target
      properties:
        ora_inizio:
          type: string
          description: Ora locale di inizio nel `fuso_orario` configurato, inclusiva. Sono accettati `HH:mm` e `HH:mm:ss`.
          pattern: '^(?:[01][0-9]|2[0-3]):[0-5][0-9](?::[0-5][0-9](?:\.[0-9]{1,9})?)?$'
          example: '06:00'
        ora_fine:
          type: string
          description: Ora locale di fine nel `fuso_orario` configurato, esclusiva; deve essere successiva a `ora_inizio`.
          pattern: '^(?:[01][0-9]|2[0-3]):[0-5][0-9](?::[0-5][0-9](?:\.[0-9]{1,9})?)?$'
          example: '08:00'
        temperatura_target:
          type: number
          format: double
          multipleOf: 0.1
          description: Temperatura target in °C con una cifra decimale.
          example: 20.5
...
    ApiError:
      type: object
      additionalProperties: false
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
          description: Istante UTC della risposta di errore.
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Payload non valido
    UnauthorizedError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          example: API key mancante o non valida
...